+++ NEWS.md
@@ -4,7 +4,7 @@
 *Development Version*
 
 #### Highlights
-
+* Add warning message to PlotPerformanceByAttribute() if multiple column names are present in both subass and attributes
 
 ## HYPEtools 1.6.2 (2024-05-24)
 *Enhancements and bug fixes*